# internal/server/spec/swagger.yaml
components:
  schemas:
    apps.AppCreateResponse:
      properties:
        compose_yaml:
          type: string
        job_queue_id:
          type: string
        name:
          type: string
        origin:
          type: string
        project_name:
          type: string
      type: object
    apps.AppRequest:
      properties:
        compose_yaml:
          type: string
        name:
          type: string
        origin:
          type: string
        project_name:
          type: string
      type: object
    apps.AppResponse:
      properties:
        id:
          type: string
        name:
          type: string
        origin:
          type: string
        project_name:
          type: string
      type: object
    auth.RefreshRequest:
      properties:
        refresh_token:
          type: string
      type: object
    auth.Token:
      properties:
        access_token:
          type: string
        expiration:
          $ref: '#/components/schemas/time.Duration'
        refresh_token:
          type: string
      type: object
    auth.UserLoginRequest:
      properties:
        email:
          type: string
        password:
          type: string
      type: object
    auth.UserResponse:
      properties:
        avatar:
          type: string
        email:
          type: string
        formatted_name:
          type: string
        username:
          type: string
      type: object
    github_com_Lakelimbo_kaeru_tools_errors.Error:
      properties:
        fields:
          additionalProperties:
            type: string
          type: object
        identifier:
          type: string
        message:
          type: string
        raw_error: {}
      type: object
    time.Duration:
      enum:
      - -9223372036854775808
      - 9223372036854775807
      - 1
      - 1000
      - 1000000
      - 1000000000
      - 60000000000
      - 3600000000000
      - -9223372036854775808
      - 9223372036854775807
      - 1
      - 1000
      - 1000000
      - 1000000000
      - 60000000000
      - 3600000000000
      type: integer
      x-enum-varnames:
      - minDuration
      - maxDuration
      - Nanosecond
      - Microsecond
      - Millisecond
      - Second
      - Minute
      - Hour
      - minDuration
      - maxDuration
      - Nanosecond
      - Microsecond
      - Millisecond
      - Second
      - Minute
      - Hour
externalDocs:
  description: ""
  url: ""
info:
  title: ""
  version: ""
openapi: 3.1.0
paths:
  /api/v1/apps:
    get:
      description: Lists all the apps
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/apps.AppResponse'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Internal Server Error
      summary: List
      tags:
      - apps
    post:
      description: Creates an app with Compose manually
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/apps.AppRequest'
                description: Data to create the app stack
                summary: request
        description: Data to create the app stack
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apps.AppCreateResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Internal Server Error
      summary: Create
      tags:
      - apps
  /api/v1/apps/{id}:
    get:
      description: Views an app
      parameters:
      - description: App ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apps.AppResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Internal Server Error
      summary: View
      tags:
      - apps
  /api/v1/auth/login:
    post:
      description: Authenticates the user
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/auth.UserLoginRequest'
                description: Login information
                summary: request
        description: Login information
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.Token'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Internal Server Error
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      description: Logs out the user
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/auth.RefreshRequest'
                description: Refresh token to be deleted
                summary: request
        description: Refresh token to be deleted
        required: true
      responses:
        "200":
          content:
            application/json: {}
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Internal Server Error
      summary: Logout
      tags:
      - auth
  /api/v1/auth/me:
    get:
      description: Basic user information
      parameters:
      - description: Access token
        in: header
        name: Authorization
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.UserResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Internal Server Error
      summary: Me
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      description: Refreshes the authentication token
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/auth.RefreshRequest'
                description: Refresh token to be swapped
                summary: request
        description: Refresh token to be swapped
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.Token'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Lakelimbo_kaeru_tools_errors.Error'
          description: Internal Server Error
      summary: Refresh
      tags:
      - auth
  /api/v1/jobs/{id}/stream:
    get:
      description: Realtime job status stream
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: OK
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Internal Server Error
      summary: Stream
      tags:
      - jobs